+++ inodaqv2/main.py
@@ -5,47 +5,10 @@ from flask import Flask, render_template, request, jsonify
 from werkzeug.wrappers.response import Response
 from click import command, option
 from inoio import errors
+from inodaqv2.components import actions
 from inodaqv2.components.extensions import conn
 
 app = Flask(__name__)
-
-
-def toggle_digital_pins(pin: str, state: bool) -> dict[str, str]:
-    pin_id = pin.split("-")[1]
-
-    command = f"dig:{pin_id}:"
-
-    if state:
-        command += "on"
-    else:
-        command += "off"
-
-    try:
-        conn.write(command)
-    except errors.InoIOTransmissionError as e:
-        message = str(e)
-    else:
-        message = conn.read()
-
-    return {"command": command, "message": message}
-
-
-def read_analog_pins() -> dict[str, int]:
-    try:
-        conn.write("aread")
-    except errors.InoIOTransmissionError as e:
-        message = str(e)
-    else:
-        message = conn.read()
-
-    _, values = message.split(";")
-    analog_to_volt = 5.0 / 1023
-
-    payload = {}
-    for u, v in enumerate(values.split(",")):
-        payload[f"A{u}"] = round(int(v) * analog_to_volt, 3)
-
-    return payload
 
 
 @app.route("/", methods=["GET", "POST"])
@@ -55,10 +18,12 @@ def dashboard() -> Union[Response, str]:
             payload = loads(request.data)
 
             if payload["action"] == "dig":
-                return jsonify(toggle_digital_pins(payload["pin"], payload["state"]))
+                return jsonify(
+                    actions.toggle_digital_pins(payload["pin"], payload["state"])
+                )
 
             if payload["action"] == "aread":
-                return jsonify(read_analog_pins())
+                return jsonify(actions.read_analog_pins())
 
     return render_template("dashboard.html")
